Move device_time to a mixin class (#314)
