+++ appium/webdriver/webdriver.py
@@ -28,6 +28,7 @@ from .extensions.activities import Activities
 from .extensions.applications import Applications
 from .extensions.clipboard import Clipboard
 from .extensions.context import Context
+from .extensions.device_time import DeviceTime
 from .extensions.images_comparison import ImagesComparison
 from .extensions.ime import IME
 from .extensions.hw_actions import HardwareActions
@@ -99,6 +100,7 @@ class WebDriver(
     Applications,
     Clipboard,
     Context,
+    DeviceTime,
     HardwareActions,
     ImagesComparison,
     IME,
@@ -522,29 +524,6 @@ class WebDriver(
         return self
 
     @property
-    def device_time(self):
-        """Returns the date and time from the device.
-        """
-        return self.execute(Command.GET_DEVICE_TIME_GET, {})['value']
-
-    def get_device_time(self, format=None):
-        """Returns the date and time from the device. (Only available since Appium 1.11.0)
-
-        :Args:
-         - format - (optional) The set of format specifiers. Read https://momentjs.com/docs/
-           to get the full list of supported datetime format specifiers.
-           If unset, return :func:`.device_time` as default format is `YYYY-MM-DDTHH:mm:ssZ`,
-           which complies to ISO-8601
-
-        :Usage:
-            self.driver.get_device_time()
-            self.driver.get_device_time("YYYY-MM-DD")
-        """
-        if format is None:
-            return self.device_time
-        return self.execute(Command.GET_DEVICE_TIME_POST, {'format': format})['value']
-
-    @property
     def battery_info(self):
         """
         Retrieves battery information for the device under test.
@@ -601,9 +580,5 @@ class WebDriver(
             ('POST', '/session/$sessionId/appium/settings')
         self.command_executor._commands[Command.LOCATION_IN_VIEW] = \
             ('GET', '/session/$sessionId/element/$id/location_in_view')
-        self.command_executor._commands[Command.GET_DEVICE_TIME_GET] = \
-            ('GET', '/session/$sessionId/appium/device/system_time')
-        self.command_executor._commands[Command.GET_DEVICE_TIME_POST] = \
-            ('POST', '/session/$sessionId/appium/device/system_time')
         self.command_executor._commands[Command.CLEAR] = \
             ('POST', '/session/$sessionId/element/$id/clear')
